Reduce minimum bag size to zero
This is not a change that should be necessary. However, node actors
have manually performed a pseudo-delete of an erroneous bag, setting
size to zero in the process. We must make this change to match the
new reality.

The admin node for the bag should have just abandoned trying to
replicate it.