Actually get the config to run
This commit includes the use of the lonely operator (&.). We do
this because even though the production configuration is not used
in dev or test, the ERB is still parsed.

This problem is unavoidable with database.yml. It may be unavoidable
with secrets.yml. It should be avoidable with all other configuration
files. From there, we can just supply the sensible defaults for the
non-production environments via ettin in config/settings.